Remove `autocomplete` and inline `evaluate`
